feat(rpc): add curated resource types for UI consumption
Hand-written API model types (Pipeline, PipelineRun, Trigger, TriggerRun)
lived in app config but are really API-layer types any consumer needs.
Exported via subpath @michelangelo-ai/rpc/resource-types to avoid name
collisions with proto-generated types. App re-exports from rpc to keep
the diff minimal.
