+++ javascript/app/config/entities/pipeline/types.ts
@@ -1,11 +1,1 @@
-export interface Pipeline {
-  metadata: {
-    name: string;
-    namespace: string;
-  };
-  spec: {
-    owner: {
-      name: string;
-    };
-  };
-}
+export type { Pipeline } from '@michelangelo-ai/rpc/resource-types';
