+++ javascript/app/config/entities/run/types.ts
@@ -1,18 +1,1 @@
-export type PipelineRun = {
-  metadata: {
-    name: string;
-    namespace: string;
-  };
-  spec: {
-    /** Populated server-side from the `x-user-name` request header, not set by the client. */
-    actor?: {
-      name: string;
-    };
-    pipeline: {
-      name: string;
-      namespace: string;
-    };
-    /** Optional human-readable description for this run. */
-    description?: string;
-  };
-};
+export type { PipelineRun } from '@michelangelo-ai/rpc/resource-types';
