+++ javascript/packages/rpc/index.ts
@@ -32,3 +32,4 @@ export * from './gen/michelangelo/api/v2/user_pb';
 
 export { request } from './request';
 export { normalizeTranscoderError } from './normalize-transcoder-error';
+
